Initial commit of Maxwell2d-0.6.4
maxwell2d-0.6.4.tar.gz
This package provides the code for programs that read a simple ASCII
configuration file that describes a distribution of dielectric
constant, and then runs the simulation. The output is either in the
form of an animated gif or a NetCDF data file.

Maxwell2D -- a 2D simulation of Maxwell's equations to demonstrate
physical phenomena associated with electromagnetic waves

Robin Hogan
r.j.hogan@reading.ac.uk
5 August 2009
http://www.met.reading.ac.uk/clouds/maxwell/ (retrieved Nov 16,2016)